# edcconnectoradapter-main/openapi/openapi-v1.yaml
openapi: 3.0.1
info:
  title: EDC Connector Adapter API
  description: |-
    APIs enable seamless communication with the EDC Connector, supporting a modular architecture that separates the responsibilities of asset/policy registration, contract negotiation and data consumption. This set of APIs facilitates secure, standardized, and traceable data exchange between providers and consumers in federated data ecosystems.

    ## Business Context

    The EDC Connector Adapter acts as an intermediary between provider-side and consumer-side applications and the EDC Connector. It enables:

    1. Registration and enrichment of assets and policies on the provider side.

    2. Discovery of data offers from provider connector catalog.

    3. Negotiation of usage contracts based on defined policies.

    4. Initiation and monitoring of data transfers.


    ### Key Actors

    - **Provider**: Publishes data assets, defines contract terms, and registers policies.

    - **Consumer**: Searches for offers, negotiates contracts, and consumes data.

    ### Typical Flow

    1. Retrieve Participant Identity

    2. Register and Enrich Assets and Policies

    3. Retrieve Asset Details

    4. Initiate Contract Negotiation

    5. Check Contract Status

    6. Initiate Data Transfer

    7. Monitor Transfer Status
  contact:
    name: Simpl Programme
    url: https://simpl-programme.ec.europa.eu/
    email: cnect-simpl@ec.europa.eu
  license:
...
  version: "1.0"
servers:
...
security:
- NoAuth: []
paths:
  /transfers:
    post:
      tags:
      - Transfer
      summary: Initiate a new transfer process
      description: Starts a data transfer process using a previously negotiated contract. This operation enables the actual movement of data from provider to consumer.
      operationId: startTransfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
            examples:
              transferRequest:
                description: transferRequest
                value:
                  contractId: 4a7f28e5-33a9-4f9e-9de9-a4f687502bce
                  dataDestination:
                    bucketName: simpl-cons
                    keyName: test-key-name
                    objectName: european_health_data.csv
                    path: folder1/
                    region: de
                    storage: s3-eu-central-1.ionoscloud.com
                    type: IonosS3
                  providerEndpoint: https://edc-provider.dev.simpl-europe.eu/protocol
                  properties:
                    resourceAddressTemplateId: TPL-1
                    key2: value2
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                Internal Server Error Example:
                  description: Internal Server Error Example
                  value:
                    type: urn:problem-type:simpl:internalServerError
                    title: Internal Server Error
                    status: 500
                    detail: Unexpected internal error
                    instance: /transfers
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Invalid payload
                    status: 400
                    detail: missing required arguments
                    issues:
                    - detail: Field 'dataDestination' must not be empty
                    instance: /transfers
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferProcessId'
              examples:
                transferProcessId:
                  description: transferProcessId
                  value:
                    transferProcessId: transfer-process-567890
...
  /transfers/{transferProcessId}:
    get:
      tags:
      - Transfer
      summary: Returns the transfer process status
      description: Returns the current status of a data transfer process. Allows consumers to monitor the progress and completion of data delivery.
      operationId: getTransferStatus
      parameters:
      - name: transferProcessId
        in: path
        description: The ID of the transfer process to retrieve the status for
        required: true
        schema:
          type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                Internal Server Error Example:
                  description: Internal Server Error Example
                  value:
                    type: urn:problem-type:simpl:internalServerError
                    title: Internal Server Error
                    status: 500
                    detail: Unexpected internal error
                    instance: /transfers/{transferProcessId}
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Invalid payload
                    status: 400
                    detail: missing required arguments
                    issues:
                    - detail: Field 'transferId' must not be empty
                    instance: /transfers/{transferProcessId}
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferProcess'
              examples:
                transferProcess:
                  description: transferProcess
                  value:
                    id: transfer-process-567890
                    state: COMPLETED
                    type: CONSUMER
                    errorDetail: 
                    dataDestination:
                      type: HttpData
                      baseUrl: https://consumer-destination.example.com
                    contractAgreementId: agreement-xyz789
        "404":
          description: NotFound
          content:
            application/problem+json:
              examples:
                Unauthorized Error Example:
                  description: Unauthorized Error Example
                  value:
                    type: urn:problem-type:simpl:unauthorized
                    title: Unauthorized
                    status: 404
                    detail: Missing or invalid Authorization header
                    instance: /transfers/{transferProcessId}
...
components:
  schemas:
    InternalServerErrorProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        additionalProperties:
          type: object
          additionalProperties:
            type: object
    BadRequestProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        issues:
          type: array
          items:
            $ref: '#/components/schemas/InputValidationIssue'
        additionalProperties:
          type: object
          additionalProperties:
            type: object
    InputObject:
      type: object
      properties:
        in:
          type: string
          enum:
          - body
          - path
          - query
          - header
        name:
          type: string
        value:
          type: object
    InputValidationIssue:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        detail:
          type: string
        in:
          type: string
          enum:
          - body
          - path
          - query
          - header
        name:
          type: string
        value:
          type: object
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/InputObject'
        additionalProperties:
          type: object
          additionalProperties:
            type: object
    JsonNode:
      type: object
    TransferRequest:
      required:
      - contractId
      - dataDestination
      - providerEndpoint
      - templateId
      type: object
      properties:
        providerEndpoint:
          minLength: 1
          type: string
        contractId:
          minLength: 1
          type: string
        templateId:
          minLength: 1
          type: string
        dataDestination:
          $ref: '#/components/schemas/JsonNode'
    TransferProcessId:
      type: object
      properties:
        transferProcessId:
          type: string
...
    TransferProcess:
      type: object
      properties:
        finalState:
          type: string
        assetId:
          type: string
        contractId:
          type: string
        correlationId:
          type: string
        transferType:
          type: string
        errorDetail:
          type: string
        type:
          type: string
        stateTimestamp:
          type: integer
          format: int64
        '@id':
          type: string
        state:
          type: string
...
  securitySchemes:
    NoAuth:
      type: apiKey
      description: Placeholder scheme used when security is not required
      name: X-No-Auth
      in: header